renamed public pages to use index.js and style.scss
