+++ src/App.js
@@ -8,10 +8,10 @@ import Login from "./pages/admin/Login/Login";
 import Register from "./pages/admin/Register/Register";
 import AdminBlog from "./pages/admin/Blog/index";
 // import Footer from "./components/admin/Footer/Footer";
-import Home from "./pages/public/Home/Home";
-import Services from "./pages/public/Services/Services";
-import SampleService from "./pages/public/Services/SampleService";
-import Leadership from "./pages/public/Leadership/Leadership";
+import Home from "./pages/public/Home";
+import Services from "./pages/public/Services";
+import SampleService from "./pages/public/Services/SpecificServices";
+import Leadership from "./pages/public/Leadership";
 import Blog from "./pages/public/Blog";
 import Contact from "./pages/public/Contact/ContactUs";
 
@@ -100,7 +100,7 @@ export default class App extends Component {
                 <Route exact path="/home" component={Home} />
                 <Route exact path="/services" component={Services} />
                 <Route exact path="/sampleservice" component={SampleService} />
-                {/* <Route exact path="/leadership" component={Leadership} /> */}
+                <Route exact path="/leadership" component={Leadership} />
                 <Route exact path="/blog" component={Blog} />
                 <Route exact path="/contact" component={Contact} />
               </Switch>
