+++ src/pages/public/Home/index.js
@@ -1,7 +1,7 @@
 import image from "../../../assets/images/rockclimber.png";
 import NavBar from "../../../components/public/NavBar";
 import Footer from '../../../components/admin/Footer/Footer';
-import "./Home.css";
+import "./style.css";
 import React, { Component } from "react";
 
 class Services extends Component {
